% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervals.R
\name{convert_10x_peak_names_to_misha_intervals}
\alias{convert_10x_peak_names_to_misha_intervals}
\title{Convert 10X-format peak names to misha intervals}
\usage{
convert_10x_peak_names_to_misha_intervals(pn, add_intervalID = FALSE)
}
\arguments{
\item{pn}{peak names as generated by 10X ATAC pipeline ("{chr}-{start}-{end}" format)}

\item{add_intervalID}{whether to add intervalID field to ouptut}
}
\value{
intervals - the peak names converted to misha-compatible intervals
}
\description{
Convert 10X-format peak names to misha intervals
}
\examples{
\dontrun{
pn <- rownames(my_scatac_mat)
my_intervals <- convert_10x_peak_names_to_misha_intervals(pn)
}
}
