% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fwrite_header}
\alias{fwrite_header}
\title{writes a data frame with a header}
\usage{
fwrite_header(data, file, header, sep = ",", ...)
}
\arguments{
\item{data}{data frame}

\item{file}{output file name}

\item{header}{string with the header}

\item{...}{
  Arguments passed on to \code{\link[tgutil:fwrite]{tgutil::fwrite}}
  \describe{
    \item{\code{}}{}
  }}
}
\description{
writes a data frame with a header
}
\examples{
\dontrun{
fwrite_header(mtcars, file = tempfile(), header = "# this is a header", sep = "\\\\t")
}

}
