% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igv.R
\name{fwrite_igv}
\alias{fwrite_igv}
\title{Write IGV file from intervals set data frame}
\usage{
fwrite_igv(
  data,
  file,
  type = NULL,
  rm_intervalID = TRUE,
  feature_name = 1,
  ...
)
}
\arguments{
\item{data}{data frame with chrom,start,end and additional data columns}

\item{file}{name of the output IGV file}

\item{type}{data type for IGV viewer. Available options are: COPY_NUMBER, GENE_EXPRESSION, CHIP, DNA_METHYLATION, ALLELE_SPECIFIC_COPY_NUMBER, LOH, RNAI. Note that if \code{type} is set other viewer options (in \code{...}) would be ignored.}

\item{rm_intervalID}{remove intervalID column from \code{data}}

\item{feature_name}{feature name for IGV viewer}

\item{...}{track parameters for IGV viewer. See: http://software.broadinstitute.org/software/igv/TrackLine.}
}
\description{
Write IGV file from intervals set data frame
}
\examples{
\dontrun{
fwrite_igv(d, "out.igv", graphType = "points")
fwrite_igv(d, "out.igv", type = "DNA_METHYLATION")
}

}
