% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucsc.R
\name{fwrite_ucsc}
\alias{fwrite_ucsc}
\title{Write USCS genome browser file from intervals set data frame}
\usage{
fwrite_ucsc(
  intervals,
  file,
  name,
  type = NULL,
  description = "",
  color = "black",
  rm_intervalID = TRUE,
  append = FALSE,
  sparse = TRUE,
  span = NULL,
  ...
)
}
\arguments{
\item{intervals}{misha intervals set. Can have an additional 'name' and 'score' fields.}

\item{file}{name of the output file}

\item{name}{track name}

\item{type}{track type. See https://genome.ucsc.edu/FAQ/FAQformat.html}

\item{description}{track description}

\item{color}{color of the track in the genome browser}

\item{rm_intervalID}{remove intervalID column from \code{intervals}}

\item{append}{append to an existing file}

\item{span}{span parameter for \code{type = "wig"}}

\item{...}{list of additional attributes such as group or priority. See: https://genome.ucsc.edu/goldenPath/help/customTrack.html}
}
\description{
Write USCS genome browser file from intervals set data frame
}
\examples{
\dontrun{
gset_genome("mm9")
intervals <- gintervals(1, c(3025716, 3052742, 3181668), c(3026216, 3053242, 3182168))
intervals$score <- c(0.2, 0.5, 1)
fwrite_ucsc(intervals, "out.ucsc", name = "clust1", type = "bedGraph", graphType = "bar", color = "red", viewLimits = "0:1", autoScale = "off")
intevals2 <- gintervals(1, c(3671488, 3903482, 3943609), c(3671988, 3903982, 3944109))
intevals2$score <- c(1, 0.2, 0.45)
fwrite_ucsc(intervals2, "out.ucsc", name = "clust2", type = "bedGraph", graphType = "bar", color = "blue", viewLimits = "0:1", autoScale = "off", append = TRUE)
}

}
