% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groot.R
\name{genome_exists}
\alias{genome_exists}
\title{Does a genome db exist at the config file}
\usage{
genome_exists(genome, params_yaml = find_params_yaml())
}
\arguments{
\item{genome}{name of the genome (e.g. hg19)}

\item{params_yaml}{path to .misha.yaml parameters file. By default, \code{misha.ext} would look for such file}
}
\description{
Does a genome db exist at the config file
}
