% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylation.R
\name{gextract_meth}
\alias{gextract_meth}
\title{gextract methylation data}
\usage{
gextract_meth(
  tracks,
  names = NULL,
  intervals = gintervals.all(),
  iterator = "intervs.global.seq_CG",
  d_expand = NULL,
  join_intervals = FALSE,
  extract_meth_calls = FALSE,
  avg_tracks = NULL,
  avg_tracks_names = NULL,
  annot_tracks = NULL,
  annot_tracks_names = NULL,
  min_cov = NULL,
  ...
)
}
\arguments{
\item{tracks}{name of methylation track suffix (without ".cov" or ".meth")}

\item{names}{sets the columns names in the returned value.}

\item{intervals}{genomic scope for which the function is applied}

\item{iterator}{track expression iterator. If "NULL" iterator is based on CpGs}

\item{d_expand}{smooth the methylation signal \code{d_expand} bp from each side (optional).}

\item{join_intervals}{add the intervals to the returned data frame (using left_join)}

\item{extract_meth_calls}{extract also methylation calls (".meth" tracks)}

\item{avg_tracks}{tracks of average methylation (for tracks that do not have ".cov" and ".meth" tracks)}

\item{avg_tracks_names}{names of tracks in \code{avg_tracks} (optional)}

\item{annot_tracks}{other tracks to extract}

\item{annot_tracks_names}{names of annotation tracks}

\item{min_cov}{minimal coverage per track. If the coverage is below this number the methylation values would be changed to NA.}

\item{...}{additional arguments to \link[misha]{gextract}}
}
\value{
data frame with a column with average methylation for each track, and another column with ".cov" suffix with every track's coverage. If \code{extract_meth_calls} is TRUE - additional columns with ".meth" suffix with number of methylation calls per track.
}
\description{
gextract methylation data
}
\seealso{
\link[misha]{gextract}
}
