% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervals.R
\name{gintervals.centers}
\alias{gintervals.centers}
\title{Calculate the centers of intervals set}
\usage{
gintervals.centers(inv)
}
\arguments{
\item{inv}{intervals set}
}
\value{
a single basepair intervals set with the center of each interval
}
\description{
Calculate the centers of intervals set
}
