% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervals.R
\name{gintervals.distance}
\alias{gintervals.distance}
\title{Calculate pairwise distance between interval sets}
\usage{
gintervals.distance(intervals1, intervals2)
}
\arguments{
\item{intervals1}{first intervals set}

\item{intervals2}{first intervals set (should be the same length as \code{intervals1})}
}
\value{
a vector containing for each i..N the distance between each interval1\link{i} and intervals2\link{i}, where N is the number
of intervals
}
\description{
Calculate pairwise distance between interval sets
}
