% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervals.R
\name{gintervals.expand}
\alias{gintervals.expand}
\title{Exapnd an intervals set}
\usage{
gintervals.expand(inv, expansion = 100)
}
\arguments{
\item{inv}{intervals set}

\item{expansion}{number of bp to add from each side}
}
\value{
intervals set expanded by \code{expansion} bp, where intervals outside of the
chromosome boundries are truncated
}
\description{
Exapnd an intervals set
}
