% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervals.R
\name{gintervals.neighbors1}
\alias{gintervals.neighbors1}
\title{Finds neighbors between two sets of intervals (and does not return conflicting column names)}
\usage{
gintervals.neighbors1(
  intervals1 = NULL,
  intervals2 = NULL,
  maxneighbors = 1,
  mindist = -1e+09,
  maxdist = 1e+09,
  na.if.notfound = TRUE,
  fields = NULL
)
}
\arguments{
\item{intervals1, intervals2}{intervals}

\item{maxneighbors}{maximal number of neighbors}

\item{mindist, maxdist}{distance range for 1D intervals}

\item{na.if.notfound}{if 'TRUE' return 'NA' interval if no matching
neighbors were found, otherwise omit the interval in the answer}

\item{fields}{select only these fields from \code{intervals2}. Note that when there are conflicting
names - the repaired name should be used, see \code{tibble::repair_names}.}
}
\value{
a data frame containing the pairs of intervals from 'intervals1', intervals from 'intervals2' (with a suffix of '1', i.e. chrom1, start1 and end1), and an additional column named 'dist' ('dist1' and 'dist2' for 2D intervals) representing the distance between the corresponding intervals.
}
\description{
Finds neighbors between two sets of intervals (and does not return conflicting column names)
}
\seealso{
\link[misha]{gintervals.neighbors}
}
