% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervals.R
\name{gintervals.normalize}
\alias{gintervals.normalize}
\title{Normalize an intervals set to be of the same length}
\usage{
gintervals.normalize(inv, size)
}
\arguments{
\item{inv}{intervals set}

\item{size}{desired length of the intervals}
}
\value{
The normalized intervals set
}
\description{
The function adds \code{floor(size / 2)} bp for each side from each
interval's center
}
