% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervals.R
\name{grandom_genome}
\alias{grandom_genome}
\title{Generate random genome intervals}
\usage{
grandom_genome(size, n, dist_from_edge = 3e+06, chromosomes = NULL)
}
\arguments{
\item{size}{The size of the intervals to generate}

\item{n}{The number of intervals to generate}

\item{dist_from_edge}{The minimum distance from the edge of the chromosome for a region to start or end(default: 3e6)}

\item{chromosomes}{The chromosomes to sample from (default: all chromosomes)}
}
\value{
A data.frame with columns chrom, start, and end
}
\description{
Generate a random genome intervals with a specified number of regions of a specified size.
}
\examples{
\dontrun{
library(misha)
gdb.init_examples()
intervals <- grandom_genome(100, 1000, dist_from_edge = 100)
head(intervals)
}

}
