% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq.R
\name{gseq.create_track}
\alias{gseq.create_track}
\title{Defunct: use \code{gtrack.create_kmer} instead}
\usage{
gseq.create_track(...)
}
\arguments{
\item{...}{formerly accepted arguments; now ignored.}
}
\description{
\code{gseq.create_track} has been removed in favour of
\code{\link[=gtrack.create_kmer]{gtrack.create_kmer()}}, which is faster (whole-genome scans in seconds),
produces a dense track, and supports multi-kmer summation
(e.g. \code{c("G", "C")} for GC content) plus an optional sliding window.
}
\details{
Migration:
\preformatted{
# Old:
gseq.create_track("CG", "seq.CG")
# New:
gtrack.create_kmer("seq.CG", "CG starts (fwd)",
                   kmer = "CG", iterator = 1, strand = 1)
}
}
\seealso{
\code{\link[=gtrack.create_kmer]{gtrack.create_kmer()}}
}
