% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{gtrack.create_dirs}
\alias{gtrack.create_dirs}
\title{Create directories needed for track creation}
\usage{
gtrack.create_dirs(track, showWarnings = TRUE, mode = "0777")
}
\arguments{
\item{track}{track name}

\item{showWarnings}{see 'dir.create'}

\item{mode}{see 'dir.create'}
}
\description{
Create directories needed for track creation
}
\examples{
\dontrun{
gtrack.create_dirs("proj.sample.my_track")
}

}
