% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track.R
\name{gtrack.marginal_sum}
\alias{gtrack.marginal_sum}
\title{Create a dense track that is a sum of misha tracks}
\usage{
gtrack.marginal_sum(
  out_track,
  tracks = NULL,
  track_regex = NULL,
  description = NULL,
  iterator = 10
)
}
\arguments{
\item{out_track}{name of the track to create}

\item{tracks}{vector of tracks to sum}

\item{track_regex}{regular expression of the tracks to sum. You should provide either this or \code{tracks}.}

\item{description}{a character string description}

\item{iterator}{track expression iterator. If 'NULL' iterator is determined implicitly based on track expression.}
}
\value{
None.
}
\description{
The function creates a dense track that is a sum of misha tracks, by first creating a virtual track for each track with a "sum" function, and then running \code{gtrack.create} with a track expression that uses \code{psum} to sum the tracks.
}
\examples{
\dontrun{
gset_genome("mm10")
gtrack.marginal_sum("wt_gastru.marginal", track_regex = "wt_gastru")
gtrack.marginal_sum("wt_gastru.marginal", tracks = c("wt_gastru.mc1", "wt_gastru.mc2"))
}

}
