% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtrack-array.R
\name{gvtrack.array_extract}
\alias{gvtrack.array_extract}
\title{Extract a virtual track on each array.track column}
\usage{
gvtrack.array_extract(
  track,
  intervals,
  iterator,
  func = NULL,
  slice = NULL,
  ...
)
}
\arguments{
\item{track}{array track}

\item{intervals}{genomic scope for which the function is applied}

\item{iterator}{track expression iterator. If 'NULL' iterator is determined
implicitly based on track expressions.}

\item{func}{function to apply to each column. See \code{misha::gvtrack.create}}

\item{slice}{a vector of column names or column indices or 'NULL'}

\item{...}{track expression}
}
\description{
Extract a virtual track on each array.track column
}
