% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mat_to_intervs}
\alias{mat_to_intervs}
\title{Transform a matrix to intervals set (deprecated)}
\usage{
mat_to_intervs(mat)
}
\value{
see \code{\link[misha:gintervals.from_mat]{misha::gintervals.from_mat()}}
}
\description{
Deprecated. Use \code{\link[misha:gintervals.from_mat]{misha::gintervals.from_mat()}} instead.
}
