% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misha.ext-package.R
\docType{package}
\name{misha.ext-package}
\alias{misha.ext}
\alias{misha.ext-package}
\title{misha.ext: A home to functions that extend the misha package}
\description{
This package provides additional functions that extend the functionality of the misha package.
}
\author{
\strong{Maintainer}: Aviezer Lifshitz \email{aviezer.lifshitz@weizmann.ac.il}

}
\keyword{internal}
