% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervals.R
\name{convert_10x_peak_names_to_misha_intervals}
\alias{convert_10x_peak_names_to_misha_intervals}
\title{Convert 10X-format peak names to misha intervals}
\usage{
convert_10x_peak_names_to_misha_intervals(pn)
}
\arguments{
\item{pn}{peak names as generated by 10X ATAC pipeline ("{chr}-{start}-{end}" format)}
}
\description{
Convert 10X-format peak names to misha intervals
}
