% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fwrite_header}
\alias{fwrite_header}
\title{writes a data frame with a header}
\usage{
fwrite_header(
  data,
  file,
  header,
  sep = ",",
  col_names = FALSE,
  append = FALSE,
  scipen = 20,
  ...
)
}
\arguments{
\item{data}{data frame}

\item{file}{output file name}

\item{header}{string with the header}

\item{sep}{The separator between columns. Default is \code{","}.}

\item{col_names}{write column names}

\item{append}{append to an existing file}

\item{scipen}{ \code{integer} In terms of printing width, how much of a bias should there be towards printing whole numbers rather than scientific notation? See Details. }

\item{...}{
  Arguments passed on to \code{\link[tgutil:fwrite]{tgutil::fwrite}}
  \describe{
    \item{\code{}}{}
  }}
}
\description{
writes a data frame with a header
}
\examples{
\dontrun{
fwrite_header(mtcars, file = tempfile(), header = "# this is a header", sep = "\\\\t")
}

}
