% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucsc.R
\name{fwrite_ucsc}
\alias{fwrite_ucsc}
\title{Write USCS genome browser file from intervals set data frame}
\usage{
fwrite_ucsc(
  intervals,
  file,
  name,
  type = NULL,
  description = "",
  color = "black",
  rm_intervalID = TRUE,
  append = FALSE,
  sparse = TRUE,
  span = NULL,
  ...
)
}
\arguments{
\item{intervals}{misha intervals set. Can have an additional 'name' and 'score' fields.}

\item{file}{name of the output file}

\item{name}{track name}

\item{type}{track type. See https://genome.ucsc.edu/FAQ/FAQformat.html}

\item{description}{track description}

\item{color}{color of the track in the genome browser}

\item{rm_intervalID}{remove intervalID column from \code{intervals}}

\item{append}{append to an existing file}

\item{span}{span parameter for \code{type = "wig"}}

\item{...}{list of additional attributes such as group or priority. See: https://genome.ucsc.edu/goldenPath/help/customTrack.html}
}
\description{
Write USCS genome browser file from intervals set data frame
}
\examples{
\dontrun{
fwrite_ucsc(d, "out.ucsc", type = "broadPeak", color = "red")
fwrite_ucsc(d1, "out.ucsc", type = "broadPeak", color = "blue", append = TRUE)
fwrite_ucsc(d1, "out.ucsc", type = "broadPeak", color = "green", append = TRUE, list(visibility = 3, offset = 15))
}

}
