% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq.R
\name{gseq.create_track}
\alias{gseq.create_track}
\title{Create sequence tracks for misha database}
\usage{
gseq.create_track(
  s,
  track,
  strand = 1,
  res = NULL,
  intervals = gintervals.all()
)
}
\arguments{
\item{s}{sequence to look for. can be a regex.}

\item{track}{name of the track to create}

\item{strand}{strand of the sequences}

\item{intervals}{intervals set}
}
\description{
Create tracks based on the presence of specific sequences.
}
\examples{
\dontrun{
gseq.create_track("C", "seq.C")
gseq.create_track("T", "seq.T")
gseq.create_track("CG", "seq.CG") # genome CpGs
gseq.create_track("[GC]", "seq.G_or_C") # C or G (for GC content calculations)
}
}
