% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groot.R
\name{gset_genome}
\alias{gset_genome}
\title{Set misha root based on genome name}
\usage{
gset_genome(genome, params_yaml = find_params_yaml())
}
\arguments{
\item{genome}{name of the genome (e.g. hg19)}

\item{params_yaml}{path to .misha.yaml parameters file}
}
\description{
Set misha root based on genome name
}
\examples{
\dontrun{
 gset_genome("hg19")
}

}
