% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misha.ext-package.R
\docType{package}
\name{misha.ext-package}
\alias{misha.ext}
\alias{misha.ext-package}
\title{misha.ext: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Aviezer Lifshitz \email{aviezer.lifshitz@weizmann.ac.il}

}
\keyword{internal}
